minimal install tested for jwst skope
