+++ setup.cfg
@@ -4,7 +4,7 @@ version = 0.4.1
 author = Ru Keïn
 author_email = rkein@stsci.edu
 license = MIT
-license_file = LICENSE
+license_files = LICENSE
 url = https://github.com/spacetelescope/spacekit
 description = Astronomical Data Science and Machine Learning Toolkit
 long_description = file: README.md
@@ -37,39 +37,34 @@ setup_requires =
 install_requires =
     tensorflow>=2.7
     astropy
-    astroquery
     boto3
     matplotlib<4
     numpy>=1.19
     pandas
-    progressbar
     scikit-learn
-    tables
-    tqdm
 
 [options.extras_require]
 test =
-    pytest
-docs =
-    sphinx
-    pydot
-dev =
-    black
-    flake8
     pytest
     pytest-cov
     pytest-profiling
     tox
     bandit
     moto
+    black
+    flake8
+docs =
+    sphinx
+    pydot
+    graphviz
+x =
+    astroquery
     pydot
     graphviz
     plotly
-viz =
-    pydot
-    graphviz
-    plotly
-
+    progressbar
+    tables
+    tqdm
 
 [options.package_data]
 spacekit = 
@@ -79,7 +74,6 @@ spacekit =
 testpaths = spacekit/tests
 doctest_plus = enabled
 text_file_format = md
-# addopts = --doctest-rst
 
 [coverage:run]
 omit =
