+++ spacekit/extractor/scrape.py
@@ -10,8 +10,6 @@ import json
 import csv
 from zipfile import ZipFile
 from astropy.io import fits, ascii
-from astroquery.mast import Observations
-from progressbar import ProgressBar
 from botocore.config import Config
 from decimal import Decimal
 from boto3.dynamodb.conditions import Attr
@@ -821,173 +819,6 @@ class SvmFitsScraper(FitsScraper):
 
     def scrape_fits(self):
         self.df = super().scrape_drizzle_fits()
-
-
-class MastScraper:
-    """Class for scraping metadata from MAST (Mikulsky Archive for Space Telescopes) via ``astroquery``. Current functionality for this class is limited to extracting the `target_classification` values of HAP targets from the archive. An example of a target classification is "GALAXY" - an alphanumeric categorization of an image product/.fits file. Note - the files themselves are not downloaded, just this specific metadata listed in the online archive database. For downloading MAST science files, use the ``spacekit.extractor.radio`` module. The search parameter values needed for locating a HAP product on MAST can be extracted from the fits science extension headers using the ``astropy`` library. See the ``spacekit.preprocessor.scrub`` api for an example (or the astropy documentation)."""
-
-    def __init__(
-        self, df, trg_col="targname", ra_col="ra_targ", dec_col="dec_targ", **log_kws
-    ):
-        """Instantiates a spacekit.extractor.scrape.MastScraper object.
-
-        Parameters
-        ----------
-        df : dataframe
-            dataset containing the requisite search parameter values (kwargs for this class)
-        trg_col : str, optional
-            name of the column containing the image target names, by default "targname"
-        ra_col : str, optional
-            name of the column containing the target's right ascension values, by default "ra_targ"
-        dec_col : str, optional
-            name of the column containing the target's right ascension values, by default "dec_targ"
-        """
-        self.__name__ = "MastScraper"
-        self.log = Logger(self.__name__, **log_kws).setup_logger()
-        self.df = df
-        self.trg_col = trg_col
-        self.ra_col = ra_col
-        self.dec_col = dec_col
-        self.targets = self.df[self.trg_col].unique()
-        self.targ_any = self.df.loc[df[self.trg_col] == "ANY"][
-            [self.ra_col, self.dec_col]
-        ]
-        self.target_categories = {}
-        self.other_cat = {}
-        self.categories = {}
-
-    def scrape_mast(self):
-        """Main calling function to scrape MAST
-
-        Returns
-        -------
-        dataframe
-            updated dataset with target classification categorical data added for each observation.
-        """
-        self.target_categories = self.scrape_target_categories()
-        self.other_cat = self.scrape_other_targets()
-        self.df = self.combine_categories()
-        return self.df
-
-    def scrape_target_categories(self):
-        """Scrapes the "target_classification" for each observation (dataframe rows) from MAST using ``astroquery`` and the target name. For observations where the target classification is not found (or is blank), the ``scrape_other_targets`` method will be called using a broader set of search parameters (``ra_targ`` and ``dec_targ``).
-
-        Returns
-        -------
-        dictionary
-            target name and category key-value pairs
-        """
-        self.log.info("\n*** Assigning target name categories ***")
-        self.log.info(f"\nUnique Target Names: {len(self.targets)}")
-        bar = ProgressBar().start()
-        for x, targ in zip(bar(range(len(self.targets))), self.targets):
-            if targ != "ANY":
-                obs = Observations.query_criteria(target_name=targ)
-                cat = obs[np.where(obs["target_classification"])][
-                    "target_classification"
-                ]
-                if len(cat) > 0:
-                    self.target_categories[targ] = cat[0]
-                else:
-                    self.target_categories[targ] = "None"
-            bar.update(x + 1)
-        bar.finish()
-        return self.target_categories
-
-    def backup_search(self, targ):
-        self.targ_any[targ] = self.df.loc[self.df[self.trg_col] == targ][
-            [self.ra_col, self.dec_col]
-        ]
-
-    def scrape_other_targets(self):
-        """Scrapes MAST for remaining target classifications that could not be identified using target name. This method instead uses a broader set of query parameters: the ``ra_targ`` and ``dec_targ`` coordinates along with the dataset's proposal ID. If multiple datasets are found to match, the first of these containing a target_classification value will be used.
-
-        Returns
-        -------
-        dict
-            secondary set of remaining key-value pairs (target names and scraped categories)
-        """
-        self.other_cat = {}
-        if len(self.targ_any) > 0:
-            self.log.info(f"Other targets (ANY): {len(self.targ_any)}")
-            bar = ProgressBar().start()
-            for x, (idx, row) in zip(
-                bar(range(len(self.targ_any))), self.targ_any.iterrows()
-            ):
-                self.other_cat[idx] = {}
-                propid = str(idx).split("_")[1]
-                ra, dec = row[self.ra_col], row[self.dec_col]
-                obs = Observations.query_criteria(
-                    proposal_id=propid, s_ra=[ra, ra + 0.1], s_dec=[dec, dec + 0.1]
-                )
-                cat = obs[np.where(obs["target_classification"])][
-                    "target_classification"
-                ]
-                if len(cat) > 0:
-                    self.other_cat[idx] = cat[0]
-                else:
-                    self.other_cat[idx] = "None"
-                bar.update(x)
-            bar.finish()
-        return self.other_cat
-
-    def combine_categories(self):
-        """Combines the two dictionaries (``target_categories`` and ``other_cat``) and inserts back into the original dataframe as a new column named ``category``.
-
-        Returns
-        -------
-        dataframe
-            copy of original dataset with new "category" column data appended
-        """
-        for k, v in self.target_categories.items():
-            idx = self.df.loc[self.df[self.trg_col] == k].index
-            for i in idx:
-                self.categories[i] = v
-        self.categories.update(self.other_cat)
-        cat = pd.DataFrame.from_dict(
-            self.categories, orient="index", columns=["category"]
-        )
-        self.log.info("\nTarget Categories Assigned.")
-        self.log.info(cat["category"].value_counts())
-        self.df = self.df.join(cat, how="left")
-        return self.df
-
-    def search_by_obs_id(self, index):
-        # obs_id = 'hst_10403_29_acs_sbc_total_j96029'
-        obs_id = "_".join(index.split("_")[:6])
-        prop_id = index.split("_")[1]
-        obs = Observations.query_criteria(proposal_id=prop_id, obs_id=obs_id)
-        s_ra = obs[np.where(obs["obs_id"] == obs_id)]["s_ra"]
-        s_dec = obs[np.where(obs["obs_id"] == obs_id)]["s_dec"]
-        if len(s_ra) > 0:
-            ra = s_ra[0]
-        elif len(obs) > 0:
-            ra = obs[0]["s_ra"][0]
-        else:
-            ra = 0
-        if len(s_dec) > 0:
-            dec = s_dec[0]
-        elif len(obs) > 0:
-            dec = obs[0]["s_dec"][0]
-        else:
-            dec = 0
-        if ra != 0:
-            obs = Observations.query_criteria(
-                proposal_id=prop_id, s_ra=[ra, ra + 0.1], s_dec=[dec, dec + 0.1]
-            )
-            targname = obs[np.where(obs["target_name"])]["target_name"]
-            if len(targname) > 0:
-                targ = targname[0]
-            else:
-                targ = "ANY"
-            category = obs[np.where(obs["target_classification"])][
-                "target_classification"
-            ]
-            if len(category) > 0:
-                cat = category[0]
-            else:
-                cat = "None"
-        return ra, dec, targ, cat
 
 
 class JsonScraper(FileScraper):
